confirmation timeou tests

+++ components/confirmation.js
@@ -3,22 +3,40 @@ const Extra = require('telegraf/extra')
 const Markup = require('telegraf/markup')
 
 const close = (ctx) => {
-    ctx.deleteMessage(ctx.update.callback_query.message.message_id, ctx.update.callback_query.message.chat.id)
+    if (typeof ctx !== 'undefined') {
+        ctx.deleteMessage(ctx.update.callback_query.message.message_id, ctx.update.callback_query.message.chat.id)
+    }
 }
 const showCallbackMessage = (ctx, text) => {
-    ctx.answerCbQuery(text)
+    if (typeof ctx !== 'undefined') {
+        ctx.answerCbQuery(text)
+    }
+}
+const runCallback = (callBack, result) => {
+    if (typeof callBack === 'function') callBack(result)
 }
 exports.show = (ctx, question, callBack, errCallBack) => {
     telegram.bot().action('confirmation-component-ok', (ctx) => {
         close(ctx)
-        if (typeof callBack === 'function') callBack(true)
+        runCallback(callBack, true)
         //showCallbackMessage(ctx, `Oh, ${ctx.match[0]}! Great choice`)
     })   
     telegram.bot().action('confirmation-component-cancel', (ctx) => {
         close(ctx)
-        if (typeof errCallBack === 'function') errCallBack(false)
+        runCallback(errCallBack, false)
         //showCallbackMessage(ctx, `Oh, ${ctx.match[0]}! Great choice`)
-    })      
+    })  
+    let timeout = 1
+    if (process.env.NODE_ENV) {
+        timeout = 10000 
+    }
+    setTimeout(
+        function() {
+            close(ctx)
+            runCallback(errCallBack, false)
+        },
+        timeout
+    )
     if (typeof ctx !== 'undefined') {
         ctx.reply(question, Extra.markup(
             Markup.inlineKeyboard([
